+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Pipeline ID and Name now recorded in logs for Data Extractions
+- Added support for viewing extraction logs in tree form (for a given ExtractionConfiguration)
 
 ## [4.1.9] - 2020-09-17